requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
prettytable==2.1.0
pyTelegramBotAPI
python-dateutil==2.8.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
today==0.2
trafaret==2.1.0
trafaret-config==2.0.2
urllib3==1.26.3
wcwidth==0.2.5
rollbar==0.15.2

python-dotenv~=0.17.0
